+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -324,6 +324,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed duplicate column labels in LightEditor's light tab
 - Fixed white and dark flashes on scenes with very high or very low exposure when Automatic Exposure is being used.
 - Fixed an issue where passing a null ProfilingSampler would cause a null ref exception.
+- Fixed memory leak in Sky when in matcap mode.
 
 ### Changed
 - Color buffer pyramid is not allocated anymore if neither refraction nor distortion are enabled